Move information about the installation from a dedicated widget into the About widget
