+++ .changeset/nervous-rice-search.md
@@ -1,0 +1,5 @@
+---
+'@giantswarm/backstage-plugin-gs': patch
+---
+
+On the cluster details page, move the information about the installation from a dedicated widget into the About widget.
