--- plugins/gs/src/components/clusters/cluster-details/ClusterOverview/ClusterInstallationCard/index.ts
@@ -1,1 +1,0 @@
-export { ClusterInstallationCard } from './ClusterInstallationCard';
