+++ plugins/gs/src/components/clusters/cluster-details/ClusterOverview/ClusterOverview.tsx
@@ -5,14 +5,9 @@ import { ErrorsProvider } from '../../../Errors';
 import { ClusterAccessCard } from './ClusterAccessCard';
 import { ClusterPolicyComplianceCard } from './ClusterPolicyComplianceCard';
 import { ClusterLabelsCard } from './ClusterLabelsCard';
-import { ClusterInstallationCard } from './ClusterInstallationCard';
-import { useCurrentCluster } from '../../ClusterDetailsPage/useCurrentCluster';
-import { isManagementCluster } from '@giantswarm/backstage-plugin-gs-common';
 import { ClusterToolsCard } from './ClusterToolsCard';
 
 export const ClusterOverview = () => {
-  const { cluster, installationName } = useCurrentCluster();
-
   return (
     <ErrorsProvider>
       <Grid container spacing={3} alignItems="stretch">
@@ -31,11 +26,6 @@ export const ClusterOverview = () => {
         </Grid>
         <Grid item md={6} xs={12}>
           <Grid item container spacing={3}>
-            {!isManagementCluster(cluster, installationName) && (
-              <Grid item xs={12}>
-                <ClusterInstallationCard />
-              </Grid>
-            )}
             <Grid item xs={12}>
               <ClusterToolsCard />
             </Grid>
